Fix race condition in useFeatures fetchMore
When a fetchMore request is still in flight and a change is made to the base
query the updateQuery method should not include the resolved fetchMore results
containing old filters.

There is no way to currently abort fetchMore requests, thus we resort to
a deep comparison of the used variables and relevant state variables.